refine and more tests
