+++ pkg/infoschema/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "error.go",
         "infoschema.go",
         "infoschema_v2.go",
+        "interface.go",
         "metric_table_def.go",
         "metrics_schema.go",
         "tables.go",
