+++ pkg/infoschema/infoschema_v2_test.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/pingcap/tidb/pkg/infoschema/internal"
 	"github.com/pingcap/tidb/pkg/kv"
+	"github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/stretchr/testify/require"
 )
 
@@ -29,11 +30,14 @@ func TestV2Basic(t *testing.T) {
 	}()
 	is := NewInfoSchemaV2(r, NewData())
 
-	dbInfo := internal.MockDBInfo(t, r.Store(), "testDB")
+	schemaName := model.NewCIStr("testDB")
+	tableName := model.NewCIStr("test")
+
+	dbInfo := internal.MockDBInfo(t, r.Store(), schemaName.O)
 	is.Data.addDB(1, dbInfo)
 	internal.AddDB(t, r.Store(), dbInfo)
-	tblInfo := internal.MockTableInfo(t, r.Store(), "test")
-	is.Data.add(tableItem{"testDB", dbInfo.ID, "test", tblInfo.ID, 2}, internal.MockTable(t, r.Store(), tblInfo))
+	tblInfo := internal.MockTableInfo(t, r.Store(), tableName.O)
+	is.Data.add(tableItem{schemaName.L, dbInfo.ID, tableName.L, tblInfo.ID, 2}, internal.MockTable(t, r.Store(), tblInfo))
 	internal.AddTable(t, r.Store(), dbInfo, tblInfo)
 	require.Equal(t, 1, len(is.AllSchemas()))
 	require.Equal(t, 0, len(is.SchemaTables(is.AllSchemas()[0].Name)))
@@ -41,5 +45,29 @@ func TestV2Basic(t *testing.T) {
 	require.NoError(t, err)
 	is.schemaVersion = 2
 	is.ts = ver.Ver
+	require.Equal(t, 1, len(is.AllSchemas()))
 	require.Equal(t, 1, len(is.SchemaTables(is.AllSchemas()[0].Name)))
+
+	getDBInfo, ok := is.SchemaByName(schemaName)
+	require.True(t, ok)
+	require.Equal(t, dbInfo, getDBInfo)
+	require.True(t, is.SchemaExists(schemaName))
+
+	getTableInfo, err := is.TableByName(schemaName, tableName)
+	require.NoError(t, err)
+	require.NotNil(t, getTableInfo)
+	require.True(t, is.TableExists(schemaName, tableName))
+
+	getDBInfo, ok = is.SchemaByID(dbInfo.ID)
+	require.True(t, ok)
+	require.Equal(t, dbInfo, getDBInfo)
+
+	getTableInfo, ok = is.TableByID(tblInfo.ID)
+	require.True(t, ok)
+	require.NotNil(t, getTableInfo)
+
+	require.Equal(t, 2, is.SchemaMetaVersion())
+	// TODO: support FindTableByPartitionID.
 }
+
+// TODO: test misc
